api/requirements.txt
django==3.2
black==23.3.0
psycopg2==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
